@@ -3,45 +3,38 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@ledgerhq/hw-transport-webhid": "^5.51.1",
-    "@ledgerhq/hw-transport-webusb": "^5.34.0",
-    "@material-ui/core": "^4.11.2",
-    "@material-ui/icons": "^4.11.2",
-    "@project-serum/anchor": "^0.7.0",
+    "@ledgerhq/hw-transport-webhid": "^6.11.2",
+    "@ledgerhq/hw-transport-webusb": "^6.11.2",
+    "@project-serum/anchor": "^0.18.2",
     "@project-serum/serum": "^0.13.24",
+    "@solana/buffer-layout": "^3.0.0",
     "@solana/spl-token-registry": "^0.2.273",
-    "@solana/web3.js": "1.12.0",
-    "@testing-library/jest-dom": "^5.11.6",
-    "@testing-library/react": "^11.2.2",
-    "@testing-library/user-event": "^12.6.0",
-    "@types/bn.js": "^4.11.6",
-    "@types/chrome": "^0.0.134",
-    "@types/jest": "^26.0.14",
-    "@types/node": "^14.11.4",
+    "@solana/web3.js": "1.31.0",
+    "@types/bn.js": "^5.1.0",
+    "@types/chrome": "^0.0.164",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
-    "bip32": "^2.0.5",
+    "bip32": "^3.0.1",
     "bip39": "^3.0.3",
     "bn.js": "^5.1.2",
     "bs58": "^4.0.1",
-    "buffer-layout": "^1.2.0",
     "clipboard-copy": "^4.0.1",
     "dayjs": "^1.10.4",
     "ed25519-hd-key": "^1.2.0",
-    "immutable-tuple": "^0.4.10",
-    "mdi-material-ui": "^6.21.0",
-    "notistack": "^0.6.1",
-    "qrcode.react": "^1.0.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
-    "react-helmet": "^6.1.0",
     "react-hot-loader": "^4.13.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
-    "styled-components": "^5.2.1",
-    "tweetnacl": "^1.0.3",
-    "typescript": "4.1.3",
-    "web3": "^1.2.11"
+    "react-router-dom": "^6.0.2",
+    "styled-components": "^5.2.1"
+  },
+  "devDependencies": {
+    "cross-env": "^7.0.3",
+    "gh-pages": "^3.1.0",
+    "git-format-staged": "^2.1.0",
+    "husky": "^7.0.4",
+    "prettier": "^2.2.1",
+    "react-scripts": "4.0.3",
+    "typescript": "4.5.2"
   },
   "scripts": {
     "predeploy": "git pull --ff-only && yarn && yarn build",
@@ -91,18 +84,5 @@
     "hooks": {
       "pre-commit": "git-format-staged -f 'prettier --stdin-filepath \"{}\"' src/*.js"
     }
-  },
-  "devDependencies": {
-    "cross-env": "^5.2.0",
-    "gh-pages": "^3.1.0",
-    "git-format-staged": "^2.1.0",
-    "husky": "^4.3.6",
-    "prettier": "^2.2.1"
-  },
-  "resolutions": {
-    "elliptic": "6.5.4",
-    "immer": "8.0.1",
-    "node-forge": "0.10.0",
-    "ini": "1.3.6"
   }
 }
