@@ -1,5 +1,9 @@
 import React, { lazy, Suspense, useMemo } from 'react';
-import { BrowserRouter, Route, Switch, Redirect } from 'react-router-dom';
+import {
+  BrowserRouter,
+  Routes,
+  Route,
+} from "react-router-dom";
 import { Container, Wrapper } from './components/CommonStyledComponents';
 
 const Welcome = lazy(() => import('./routes/Welcome'));
@@ -46,8 +50,8 @@ export const Pages = () => {
 
   return (
     <>
-      <Switch>
-        <Route path="/welcome" component={Welcome} />
+      <Routes>
+        <Route path="/" element={<Welcome />} />
 
         {/* popup if connecting from dex UI */}
         {/* {window.opener && !!wallet && <Redirect from="/" to="/connect_popup" />} */}
@@ -56,8 +60,7 @@ export const Pages = () => {
         {/* {!!wallet && <Redirect from="/" to="/wallet" />} */}
 
         {/* if have mnemonic in localstorage - login, otherwise - restore/import/create */}
-        <Redirect from="/" to="/welcome" />
-      </Switch>
+      </Routes>
     </>
   );
 };
