@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { hot } from 'react-hot-loader';
+import { hot } from 'react-hot-loader/root';
+
 import App from './App.tsx';
 import * as serviceWorker from './serviceWorker';
 
@@ -12,7 +13,7 @@ const render = () =>
     document.getElementById('root'),
   );
 
-render(hot(module)(App));
+render(hot(App));
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
